refactor: move `getAuthedUser` to separate file
